requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.2
Pillow==8.3.2
tensorflow==2.15.0
scipy>=1.7.0
matplotlib>=3.4.0
pandas>=1.3.0
seaborn>=0.11.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
gunicorn>=20.1.0
tensorboard>=2.8.0
mlflow>=1.20.0
h5py>=3.6.0
pydot>=1.4.2
pathlib>=1.0.1
logging>=0.5.1.2
